Add ranking to player object
Before I was sorting the player ranking based off the index but I decided to add
a ranking property to the player object.

This may come at the cost of some performance, but this will be necessary for
displaying a paginated table of players.
It also removes the need for custom sorting functions which is nice.

Without this change the player rankings would be broken on that table.