Create Modal view for when no data is available
This is a Modal fallback view which only displays when there is no data
available.

This should never really happen but on the off chance it does we now have a
fallback for it.

As a result of this new component I've made a Modal folder where the Modal
component now lives.