Resolve eslint errors from broken VSCode extension
I hate making a commit like this instead of rebasing and redoing the previous
commits but it's late and I'm tired on a Sunday.

I was not getting errors during development but was getting them during the
build process. This was due to the VScode extension being out of date.